@@ -37,7 +37,6 @@
     "bootstrap": "^4.1.3",
     "chart.js": "^2.7.3",
     "countdown": "^2.6.0",
-    "cryptr": "^4.0.2",
     "css-loader": "^1.0.0",
     "highlight.js": "^9.15.6",
     "i": "^0.3.6",
@@ -56,12 +55,13 @@
     "npm": "^6.9.0",
     "poly2tri": "^1.5.0",
     "random-string": "^0.2.0",
-    "react": "^16.5.2",
+    "react": "^16.8.6",
     "react-apexcharts": "^1.3.3",
     "react-bootstrap-table-next": "^1.2.0",
     "react-bootstrap-table2-editor": "^1.1.0",
     "react-chartjs-2": "^2.7.4",
     "react-confirm-alert": "^2.0.7",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-dom": "^16.5.2",
     "react-full-screen": "^0.2.4",
     "react-grid-layout": "^0.16.6",
@@ -78,11 +78,11 @@
     "react-router-dom": "^4.3.1",
     "react-simple-code-editor": "^0.6.1",
     "react-svg-gauge": "^1.0.10",
+    "react-swipeable": "^5.3.0",
     "react-syntax-highlighter": "^8.0.1",
     "react-three-renderer": "^3.2.4",
     "react-toastify": "^5.0.0-rc.3",
     "sass-loader": "^7.1.0",
-    "scrypt": "^6.0.3",
     "socket.io-client": "^2.1.1",
     "style-loader": "^0.23.0",
     "three": "^0.100.0",
