@@ -1,26 +1,17 @@
 import React, { Component } from "react";
 import ChangePassword from "../../components/changePassword";
-const Cryptr = require('cryptr');
-const cryptr = new Cryptr('prototype');
+import { ChangeUsername } from "../../components/changeUsername";
 export class SetUsername extends React.Component {
 
     state = {
         available: false,
         username: undefined,
         message: "",
-        isOpen: false
+        isOpen: false,
+        usernameSet: false
     }
 
     componentDidMount() {
-
-    }
-
-    onChange = () => {
-        return (evt) => {
-            this.setState({ username: this.cleaner(evt.target.value) }, () => {
-                this.checkUpdateUsername();
-            })
-        }
     }
 
     openModal = (origination) => {
@@ -44,35 +35,11 @@ export class SetUsername extends React.Component {
         }).catch(err => console.error(err.toString()))
     }
 
-
-
-    showButton = () => {
-        if (this.state.available == false) {
-            return (<div>not available try a different username.</div>)
-        } else {
-            return (<button onClick={this.updateUsername} className="btn-spot" style={{ float: "right" }} > SAVE</button>)
-        }
-    }
-
-    updateUsername = () => {
-        fetch('/api/v3/account/updateusername', {
-            method: 'POST',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ username: this.state.username })
-        }).then(response => response.json()).then((data) => {
-            this.props.usernameUpdated();
-        }).catch(err => console.error(err.toString()))
-    }
-
     changePassword = () => {
         this.openModal();
     }
 
     render() {
-
         if (this.props) {
             if (this.state.username == undefined) {
                 if (this.props.username) {
@@ -80,34 +47,41 @@ export class SetUsername extends React.Component {
                 }
             }
         }
-
         return (
             <div>
                 <ChangePassword account={this.props.account} isOpen={this.state.isOpen} closeModel={() => { this.setState({ isOpen: false }) }} />
-                <h3 style={{ marginTop: "5px" }}> Username </h3>
-                <p>Here you can change your public username. This must be unique across the system. It will affect your public url in the form of /u/username</p>
-                <input
-                    style={{ width: "50%" }}
-                    value={this.state.username}
-                    onChange={this.onChange()}
-                    autoFocus={true}
-                />
-                {this.showButton()}<hr></hr>
-                <h3> CHANGE PASSWORD </h3>
-                <button
-                    className="btn-spot floatLeft"
-                    style={{ float: "center" }}
-                    onClick={this.changePassword}
-                >
-                    CHANGE PASSWORD
+                <ChangeUsername account={this.props.account} isOpen={this.state.usernameSet} closeModel={() => { this.setState({ usernameSet: false }); this.props.usernameUpdated(); }} />
+                <div style={{ display: "grid", gridTemplateColumns: "45% 45% 10%" }}>
+
+                    <div>
+                        <br /><br />
+                        <h3> CHANGE USERNAME </h3><br />
+                        <button
+                            className="btn-spot floatLeft"
+                            style={{ float: "center", opacity: 0.8 }}
+                            onClick={() => { this.setState({ usernameSet: true }, () => { console.log(this.state) }) }}
+                        >
+                            CHANGE USERNAME
                                 </button>
+                    </div>
+                    <div>
+                        <br /><br />
+                        <h3> CHANGE PASSWORD </h3><br />
+                        <button
+                            className="btn-spot floatLeft"
+                            style={{ float: "center", opacity: 0.8 }}
+                            onClick={this.changePassword}
+                        >
+                            CHANGE PASSWORD
+                                </button>
+                    </div>
+                    <div>
+                        <br /><br /><br /><br /><br />
+                        <a href="/signout"><button className="btn-spot" style={{ float: "right", opacity: 0.8 }} > SIGN OUT</button></a>
+                    </div>
+                </div>
             </div>
         )
-    }
-
-    cleaner(str) {
-        var strLower = str.toLowerCase();
-        return strLower.replace(/\W/g, '');
     }
 }
 
