@@ -1,6 +1,7 @@
 import { confirmAlert } from 'react-confirm-alert';
 import React, { Component } from "react";
 import Modal from 'react-modal';
+import Media from "react-media";
 var loggedInUser = "";
 const customStyles = {
     content: {
@@ -15,7 +16,25 @@ const customStyles = {
         maxHeight: 'calc(100vh - 210px)',
         overflow: 'auto'
     },
-    //bacground of Pop up Modal on search
+    overlay: {
+        background: "rgba(27, 57, 77,0.9)",
+    }
+};
+
+const customStylesMobile = {
+    content: {
+        top: '50%',
+        left: '50%',
+        right: '50%',
+        bottom: 'auto',
+        marginRight: '-50%',
+        transform: 'translate(-50%, -50%)',
+        border: "none",
+        background: "rgba(3, 4, 5,0.6)",
+        maxHeight: 'calc(100vh - 210px)',
+        overflow: 'auto',
+        width: "100%"
+    },
     overlay: {
         background: "rgba(27, 57, 77,0.9)",
     }
@@ -103,6 +122,7 @@ export class ShareList extends Component {
     }
 
     setValues = (isOpen) => {
+
         if (isOpen == true && count == 0) {
             count++;
             if (this.props.type) {
@@ -158,7 +178,7 @@ export class ShareList extends Component {
                     newDeviceList.push("|");
                 }
             });
-            temp = newDeviceList.filter((users) => { return users !== "|" && users.email !== this.props.account.email })
+            temp = newDeviceList.filter((users) => { return users !== "|" && users.email !== this.props.account.email && users.username !== this.props.account.username })
             for (var look in this.state.shared) {
                 for (var i in temp) {
                     if (temp[i].sharekey == this.state.shared[look]) {
@@ -172,16 +192,31 @@ export class ShareList extends Component {
 
     userNameList = () => {
         try {
-            return (<div style={{ height: "20%" }}>
+            return (<div style={{ height: "200px", overflow: "auto" }}>
                 {
                     this.state.userSearched.map((user, i) => {
-                        if (user.shared == "no") {
-                            return <div id={user.email} key={i} className="commanderBgPanel commanderBgPanelClickable" style={{ display: this.state.checkboxstate }}>{user.email} <i className={user.icon} style={{ float: "right" }} onClick={(e) => this.handleActionCall(user)} /></div>
+
+                        if (window.innerWidth <= 599) {
+                            if (i < 10) {
+                                if (user.shared == "no") {
+                                    return <div id={user.email} key={i} className="commanderBgPanel commanderBgPanelClickable" style={{ display: this.state.checkboxstate }}>{user.email} <i className={user.icon} style={{ float: "right" }} onClick={(e) => this.handleActionCall(user)} /></div>
+                                }
+                                else {
+                                    return <div id={user.email} key={i} className="commanderBgPanel commanderBgPanelClickable" style={{ display: this.state.checkboxstate }}>{user.email} <div style={{ float: "right" }} onClick={(e) => this.unshare(user.sharekey)}>Revoke Sharing </div></div>
+                                }
+                            }
                         }
-                        else {
-                            return <div id={user.email} key={i} className="commanderBgPanel commanderBgPanelClickable" style={{ display: this.state.checkboxstate }}>{user.email} <div style={{ float: "right" }} onClick={(e) => this.unshare(user.sharekey)}>Revoke Sharing </div></div>
+                        else if (window.innerWidth > 599) {
+
+                            if (user.shared == "no") {
+                                return <div id={user.email} key={i} className="commanderBgPanel commanderBgPanelClickable" style={{ display: this.state.checkboxstate }}>{user.email} <i className={user.icon} style={{ float: "right" }} onClick={(e) => this.handleActionCall(user)} /></div>
+                            }
+                            else {
+                                return <div id={user.email} key={i} className="commanderBgPanel commanderBgPanelClickable" style={{ display: this.state.checkboxstate }}>{user.email} <div style={{ float: "right" }} onClick={(e) => this.unshare(user.sharekey)}>Revoke Sharing </div></div>
+                            }
                         }
                     })
+
                 }
             </div >)
         } catch (err) { }
@@ -241,62 +276,101 @@ export class ShareList extends Component {
     }
 
     shareDevice = () => {
+        var path = window.location.origin;
         if (this.props.type) {
-            {
-                this.setState({ show: "noDisplayShare" });
-                this.state.EmailsharedDevice = _.clone(this.state.SelectedUsers) //#region 
-                for (let dev in this.state.EmailsharedDevice) {
-                    fetch("/api/v3/admin/shareDevice", {
-                        method: "POST", headers: { "Accept": "application/json", "Content-Type": "application/json" },
-                        body: JSON.stringify({
-                            email: this.state.EmailsharedDevice[dev].email,
-                            text: 'Hi Device(s) have been shared with you ',
-                            html: '<p>Hi <br></br>' + this.props.account.username + ' has shared Device(s):<br></br> (' + this.props.chosen.map((user, i) => user.devid) + ')<br></br>  with you </p>',
-                            subject: 'SHARED DEVICE',
-                            dev: this.props.chosen.map((user, i) => user.devid + ''),
-                            type: "multi",
-                            chosen: this.props.chosen,
-                            person: this.props.account.username,
-                            publickey: this.state.EmailsharedDevice[dev].sharekey
-                        })
-                    }).then(response => response.json()).then(serverresponse => {
-                        this.setState({ qresponse: serverresponse.result })
-                        this.setState({ show: " " });
-                    }).catch(err => console.error(err.toString()));
-                }
-                this.setState({ SelectedUsers: [] })
-                this.setState({ isOpen: !this.state.isOpen })
+            this.setState({ show: "noDisplayShare" });
+            this.state.EmailsharedDevice = _.clone(this.state.SelectedUsers) //#region 
+            for (let dev in this.state.EmailsharedDevice) {
+                var htmlMessage =
+                    '<p>' +
+                    'Hi ' + this.state.EmailsharedDevice[dev].username +
+                    ',<br></br>' +
+                    '<a href="' + path + "/u/" + this.props.account.username + '">' + this.props.account.username + '</a>' + ' has shared devices:<br></br>' +
+                    this.props.chosen.map((user) => { return (' -  ' + '<a href="' + path + '/u/' + this.props.account.username + '/view/' + user.devid + '">' + user.devid + '</a>' + '<br/>') }) +
+                    '<br/>with you. </p>' +
+                    '</p><br/>Kind Regards,<br/>Prototyp3<br/><br/>' +
+                    '<a href="' + path + '"><img src="cid:nxtlogokkk" alt="Prototype3 Logo"/></a> <a href="https://github.com/IoT-nxt/prototype"><img src="cid:gitlogokkk" alt="Github Logo"/></a>';
+
+                fetch("/api/v3/admin/shareDevice", {
+                    method: "POST", headers: { "Accept": "application/json", "Content-Type": "application/json" },
+                    body: JSON.stringify({
+                        email: this.state.EmailsharedDevice[dev].email,
+                        text: 'Hi Device(s) have been shared with you ',
+                        subject: 'SHARED DEVICE',
+                        dev: this.props.chosen.map((user, i) => user.devid + ''),
+                        type: "multi",
+                        chosen: this.props.chosen,
+                        person: this.props.account.username,
+                        publickey: this.state.EmailsharedDevice[dev].sharekey,
+                        html: htmlMessage,
+                        attachments: [
+                            {
+                                filename: 'favicon.png',
+                                path: '../public/favicon.png',
+                                cid: "nxtlogokkk"
+                            },
+                            {
+                                filename: 'githubLogo.png',
+                                path: '../public/githubLogo.png',
+                                cid: "gitlogokkk"
+                            },
+                        ]
+                    })
+                }).then(response => response.json()).then(serverresponse => {
+                    this.setState({ qresponse: serverresponse.result })
+                    this.setState({ show: " " });
+                }).catch(err => console.error(err.toString()));
             }
+            this.setState({ SelectedUsers: [] })
+            this.setState({ isOpen: !this.state.isOpen })
+
         }
         else if (!this.props.type) {
-            {
-                this.setState({ show: "noDisplayShare" });
-                this.state.EmailsharedDevice = _.clone(this.state.SelectedUsers) //#region 
-                for (let dev in this.state.EmailsharedDevice) {
-                    for (let i in this.state.userSearched) {
-                        if (this.state.EmailsharedDevice[dev].email == this.state.userSearched[i].email) {
-                            this.state.userSearched[i].shared = "yes";
-                        }
+            this.setState({ show: "noDisplayShare" });
+            this.state.EmailsharedDevice = _.clone(this.state.SelectedUsers) //#region 
+            for (let dev in this.state.EmailsharedDevice) {
+                var htmlMessage =
+                    '<p>' +
+                    'Hi ' + this.state.EmailsharedDevice[dev].username +
+                    ',<br></br>' +
+                    '<a style={{ textDecoration: "none"}} href="' + path + "/u/" + this.props.username + '">' + this.props.account.username + '</a>' + ' has shared "' + '<a style={{ textDecoration: "none"}} href="' + path + "/u/" + this.props.username + "/view/" + this.props.devid + '">' + this.props.devid + '</a>' + '" with you.' +
+                    '</p><br/>Kind Regards,<br/>Prototyp3<br/><br/>' +
+                    '<a href="' + path + '"><img src="cid:nxtlogokkk" alt="Prototype3 Logo"/></a> <a href="https://github.com/IoT-nxt/prototype"><img src="cid:gitlogokkk" alt="Github Logo"/></a>';
+                for (let i in this.state.userSearched) {
+                    if (this.state.EmailsharedDevice[dev].email == this.state.userSearched[i].email) {
+                        this.state.userSearched[i].shared = "yes";
                     }
-                    fetch("/api/v3/admin/shareDevice", {
-                        method: "POST", headers: { "Accept": "application/json", "Content-Type": "application/json" },
-                        body: JSON.stringify({
-                            email: this.state.EmailsharedDevice[dev].email,
-                            text: 'Hi a Device was shared with you called ' + this.props.devid,
-                            html: '<p>Hi <br></br>' + this.props.username + ' has shared (' + this.props.devid + ') Device with you </p>',
-                            subject: 'SHARED DEVICE',
-                            dev: this.props.devid,
-                            person: this.props.username,
-                            publickey: this.state.EmailsharedDevice[dev].sharekey
-                        })
-                    }).then(response => response.json()).then(serverresponse => {
-                        this.setState({ qresponse: serverresponse.result })
-                        this.setState({ show: " " });
-                    }).catch(err => console.error(err.toString()));
                 }
-                this.setState({ SelectedUsers: [] })
-                this.setState({ isOpen: !this.state.isOpen })
+                fetch("/api/v3/admin/shareDevice", {
+                    method: "POST", headers: { "Accept": "application/json", "Content-Type": "application/json" },
+                    body: JSON.stringify({
+                        email: this.state.EmailsharedDevice[dev].email,
+                        text: 'Hi a Device was shared with you called ' + this.props.devid,
+                        subject: 'SHARED DEVICE',
+                        dev: this.props.devid,
+                        person: this.props.username,
+                        publickey: this.state.EmailsharedDevice[dev].sharekey,
+                        html: htmlMessage,
+                        attachments: [
+                            {
+                                filename: 'favicon.png',
+                                path: '../public/favicon.png',
+                                cid: "nxtlogokkk"
+                            },
+                            {
+                                filename: 'githubLogo.png',
+                                path: '../public/githubLogo.png',
+                                cid: "gitlogokkk"
+                            },
+                        ]
+                    })
+                }).then(response => response.json()).then(serverresponse => {
+                    this.setState({ qresponse: serverresponse.result })
+                    this.setState({ show: " " });
+                }).catch(err => console.error(err.toString()));
             }
+            this.setState({ SelectedUsers: [] })
+            this.setState({ isOpen: !this.state.isOpen })
         }
     }
 
@@ -413,12 +487,12 @@ export class ShareList extends Component {
     selectedUserCount = () => {
         if (this.state.SelectedUsers.length == 0) {
             return (
-                <div></div>
+                null
             )
         }
         else {
             return (
-                <div className="commanderBgPanel commanderBgPanelClickable"><h4 onClick={this.selected}>SELECTED USERS ({this.state.SelectedUsers.length})</h4>
+                <div className="commanderBgPanel commanderBgPanelClickable" ><h4 onClick={this.selected}>SELECTED USERS ({this.state.SelectedUsers.length})</h4>
                     {this.selectedUsers()}
                 </div>
             )
@@ -439,19 +513,39 @@ export class ShareList extends Component {
     render() {
         return (<div ><center>
             {this.setValues(this.props.isOpen)}
-            <Modal style={customStyles} isOpen={this.props.isOpen} onRequestClose={this.toggle}>
-                <i className={"fas fa-times " + this.state.show} onClick={() => { this.props.closeModel(this.state.qresponse); count = 0; }} style={{ color: "red" }}></i>
-                <center style={{ color: "white", display: this.state.checkboxstate }}>
-                    <br></br> Search For users to share  with<br></br>
-                    <div style={{ color: "white" }}><i className="fas fa-search" style={{ color: "white" }}></i> <input type="text" name="search" placeholder=" By email" onChange={this.search} /></div></center><br></br>
-                <br></br>{this.DevicePublic()}<div>
-                    {this.ShareButton()}</div><hr></hr>
-                <br></br>{this.selectedUserCount()}<div>
-                    {this.userNameList()}
-                </div>
-                <center>
-                </center>
-            </Modal>
+            <Media query="(max-width: 599px)">
+                {matches =>
+                    matches ? (
+                        <Modal style={customStylesMobile} isOpen={this.props.isOpen} onRequestClose={this.toggle}>
+                            <i className={"fas fa-times " + this.state.show} onClick={() => { this.props.closeModel(this.state.qresponse); count = 0; }} style={{ color: "red" }}></i>
+                            <center style={{ color: "white", display: this.state.checkboxstate }}>
+                                <br></br> Search For users to share  with<br></br>
+                                <div style={{ color: "white" }}><i className="fas fa-search" style={{ color: "white" }}></i> <input type="text" name="search" placeholder=" By email" onChange={this.search} /></div></center><br></br>
+                            <br></br>{this.DevicePublic()}<div>
+                                {this.ShareButton()}</div><hr></hr>
+                            <br></br>{this.selectedUserCount()}<div>
+                                {this.userNameList()}
+                            </div>
+                            <center>
+                            </center>
+                        </Modal>
+                    ) : (
+                            <Modal style={customStyles} isOpen={this.props.isOpen} onRequestClose={this.toggle}>
+                                <i className={"fas fa-times " + this.state.show} onClick={() => { this.props.closeModel(this.state.qresponse); count = 0; }} style={{ color: "red" }}></i>
+                                <center style={{ color: "white", display: this.state.checkboxstate }}>
+                                    <br></br> Search For users to share  with<br></br>
+                                    <div style={{ color: "white" }}><i className="fas fa-search" style={{ color: "white" }}></i> <input type="text" name="search" placeholder=" By email" onChange={this.search} /></div></center><br></br>
+                                <br></br>{this.DevicePublic()}<div>
+                                    {this.ShareButton()}</div><hr></hr>
+                                <br></br>{this.selectedUserCount()}<div>
+                                    {this.userNameList()}
+                                </div>
+                                <center>
+                                </center>
+                            </Modal>
+                        )
+                }
+            </Media>
         </center>
         </div >)
     }
