@@ -1,11 +1,8 @@
 import React from "react";
 
-import { TeltonikaConfigure } from "./plugin_teltonika_add.jsx"
-
 export const name = "Teltonika";
-export const AddDevice = require("../devicetype/teltonika.jsx");
 
-export class SettingsPanel extends React.Component {
+export class AddDevice extends React.PureComponent {
 
     state = {
         teltonikaServiceEnabled: false,
@@ -20,9 +17,7 @@ export class SettingsPanel extends React.Component {
         fetch("/api/v3/teltonika/info")
             .then(res => res.json())
             .then((result) => {
-                // console.log(result);
                 if (result.port) {
-                    // console.log(result.port);
                     this.setState({ port: result.port })
                 }
             }).catch((err) => {
@@ -34,31 +29,27 @@ export class SettingsPanel extends React.Component {
         fetch("/api/v3/teltonika/reqport")
             .then(res => res.json())
             .then((result) => {
-                // console.log(result);
                 if (result.port) {
                     this.setState({ port: result.port });
                 }
             }).catch((err) => {
                 console.log(err);
             })
-
-        //this.setState({ teltonikaServiceEnabled: true })
     }
 
     render() {
         return (
-            <div className="blockstyle">
-                <h4>TELTONIKA</h4>
-
-                <div className="blockstyle" style={{ marginBottom: 20, display: "none" }}>
-                    <h4>PUBLIC PORT</h4>
-                    DEFAULT PORT: 12000
-          <p>If you set Teltonika device to connect to this port devices will be visible to administrators only.
-            Any user can type in IMEI to add device to their account.</p>
+            <div className="blockstyle" style={{ background: "rgb(22, 32, 44)" }}>
+                <h3>TELTONIKA</h3>
+                <p>To add a device to Prototyp3, configure the device to forward its data to {window.location.origin}/port. Below is the different ports offered:</p>
+                <h5>PUBLIC PORT</h5>
+                DEFAULT PORT: 12000
+                    <p>If you set Teltonika device to connect to this port devices will be visible to administrators only.
+    Any user can type in IMEI to add device to their account.
+                    </p>
+                <h5>PUBLIC PORT</h5>
+                To set a private port, navigate to <span className="commanderBgPanel">Settings -> Teltonika -> Click Enable</span> to get a private port assigned to your account.
                 </div>
-
-                <TeltonikaConfigure enable={this.enable} port={this.state.port} />
-            </div>
         );
     }
 }
