@@ -7,8 +7,8 @@ export class Landing extends Component {
   render() {
     return (
 
-      <div>
 
+      <div>
         <div style={{ backgroundImage: "url(/splash.jpg)", backgroundSize: "100% auto", minHeight: "500px", backgroundPositionY: "center" }}>
           <div className="container" >
             <div className="row">
@@ -17,16 +17,14 @@ export class Landing extends Component {
                 <p>A new IoT solution with the goal of taking down the barriers to connecting devices to the cloud.</p>
               </div>
               <div className="col-sm-7">
-
               </div>
             </div>
           </div>
         </div>
-
-
-
+        <div className='wrap' style={{ zIndex: "-10" }}>
+          <div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div><div className='c'></div>
+        </div>
       </div>
-
     )
   }
 }
