@@ -4,6 +4,7 @@ import { Teltonika } from "./lib_teltonika"
 import { log } from "../../log"
 import { Plugin } from "../plugin"
 import express = require('express');
+var RedisEvent = require('redis-event');
 
 export class PluginTeltonika extends Plugin {
   db: any;
@@ -14,13 +15,29 @@ export class PluginTeltonika extends Plugin {
   minPort = 12001;
   maxPort = 12200;
 
+  isCluster:boolean = false;
+  ev:any;
+  clustersubs:string[] = [];
+
+  connections : Teltonika[] = [];
+
   constructor(config: any, app: express.Express, db: any, eventHub: events.EventEmitter) {
     super(app, db, eventHub);
     this.db = db;
     this.eventHub = eventHub;
+    
 
     log("PLUGIN", this.name, "LOADED");
     //default 12000 port should handle any device for any account
+
+    if (config.redis && process.env.pm_id) {
+      this.isCluster = true;
+      this.ev = new RedisEvent(config.redis.host, [this.name]);
+
+      this.ev.on("read", ()=>{
+        log(this.name, "CLUSTER", "READY")
+      })
+    }
 
     app.get("/api/v3/teltonika/info", (req: any, res: any) => {
       this.db[this.collection].findOne({ apikey: req.user.apikey }, (err: Error, result: any) => {
@@ -65,10 +82,32 @@ export class PluginTeltonika extends Plugin {
     })
   }
 
-  handlePacket(packet: any) {
-    // log("PLUGIN", this.name, "HANDLE PACKET TODO");
+  handlePacket(deviceState:any, packet: any, cb:any) {
+    log("PLUGIN", this.name, "HANDLE PACKET");
     // check if this device is a teltonika device
+    if (!deviceState) { console.log(this.name + " handlePacket ERROR no deviceState")}
+        
+    if ((this.isCluster) && (packet.fromCluster == undefined)) {
+      log(this.name, "CLUSTER", "EVENT handlePacket");
+      packet.fromCluster = true;
+      packet.fromId = process.env.pm_id;
+      this.ev.pub(this.name + ":" + deviceState.apikey, {
+        deviceState,
+        packet,
+        launchedAt: new Date()
+      })
+    }
+
     // send data to device if its connected to this node
+    for (var dev of this.connections) {
+      if (packet.payload.data.command) {
+        if (deviceState.devid == dev.id) {
+          if (deviceState.apikey == dev.apikey) {
+            dev.tcpwrite(packet.payload.data.command)
+          }          
+        }
+      }
+    }
   }
 
   findOpenPort(cb: Function) {
@@ -109,13 +148,48 @@ export class PluginTeltonika extends Plugin {
 
   connectPort(userPort: any) {
     var server = net.createServer((client: any) => {
-      var device = new Teltonika(client, {});
+      var device:any = new Teltonika(client, {apikey:userPort.apikey});
+
+      if (this.isCluster) {
+        var sub = this.name + ":" + userPort.apikey
+
+        if (this.clustersubs.indexOf(sub)== -1) {
+          this.clustersubs.push(sub);
+          log(this.name, "CLUSTER", "NEW SUB");
+          this.ev.on(sub, (data:any)=>{
+            if (data.packet.fromId != process.env.pm_id) {
+              this.handlePacket(data.deviceState, data.packet, ()=>{})
+            }
+          })
+        }
+      }
+
       device.on("data", (data: any) => {
+        data.data.connected = true;
+
         this.eventHub.emit("device", {
-          apikey: userPort.apikey,
+          apikey: userPort.apikey,  
           packet: { id: data.id, data: data.data, meta: { method: "teltonika" } }
         })
       })
+
+      device.on("end", () => {
+
+        this.eventHub.emit("device", {
+          apikey: userPort.apikey,
+          packet: { id: device.id, data: { connected : false}, meta: { method: "teltonika" } }
+        })
+
+        console.log("teltonika tcp disconnected")
+        for (var dev in this.connections) {
+          if (device == this.connections[dev]) {
+            console.log("found device in list")
+            this.connections.splice(parseInt(dev),1);
+          }
+        }
+      })
+
+      this.connections.push(device);
     })
 
     server.listen(userPort.port, () => {
