@@ -342,9 +342,9 @@ export class Prototype extends EventEmitter {
         })
     }
 
-    setdevicegateway(id: string, GatewayId: string, HostAddress: string, cb: any) {
+    setdevicegateway(id: string, GatewayId: string, HostAddress: string, options: any, cb: any) {
         this.state(id, (err: any, device: any) => {
-            this.plugins.iotnxt.setgatewaydevice(device.key, id, GatewayId, HostAddress, (err: Error, result: any) => {
+            this.plugins.iotnxt.setgatewaydevice(device.key, id, GatewayId, HostAddress, options, (err: Error, result: any) => {
                 cb(err, result);
             })
         })
@@ -427,6 +427,42 @@ Device Dashboard
                 if (err) cb(err);
                 else if (body) {
                     if (body.nModified == 1) {
+                        cb(null, body);
+                    } else {
+                        cb(body);
+                    }
+                }
+            });
+        })
+    }
+    /*-----------------------------------------------------------------------------------------------*/
+
+    /*
+CLEAR DEVICE DATA & WORKFLOW
+*/
+    //CLEAR DEVICE DATA
+    clearDeviceData(id: any, cb: Function) {
+        request.post(this.uri + "/api/v3/state/clear", { headers: this.headers, json: { id: id } }, (err, res, body) => {
+            if (err) cb(err);
+            else if (body) {
+                if (body.nModified == 1) {
+                    cb(null, body);
+                } else {
+                    cb(body);
+                }
+            }
+        });
+    }
+
+    //CEAR DEVICE DATA & HISTORY
+    clearDevDataHist(id: any, cb: Function) {
+        this.state(id, (err: any, device: any) => {
+            var devices = []
+            devices.push(device)
+            request.post(this.uri + "/api/v3/state/clear", { headers: this.headers, json: { id: devices, type: "multi", clearhistory: true } }, (err, res, body) => {
+                if (err) cb(err);
+                else if (body) {
+                    if (body.ok == 1) {
                         cb(null, body);
                     } else {
                         cb(body);
